dinov2-rhd/requirements.txt
torch
torchvision
google-cloud-storage 
pandas 
pillow
numpy 
matplotlib 
seaborn 
scikit-lean
umap-learn 
tqdm